Rename BaseCharge to Charge and Charge to ChargeCreator
The existing Charge class' sole purpose is to create one of the subclasses of BaseCharge. It makes sense to rename it to ChargeCreator. This opens up the name Charge to be used instead of BaseCharge.
