+++ src/backend/expungeservice/models/charge_types/felony_class_b.py
@@ -1,8 +1,8 @@
-from expungeservice.models.charge_types.base_charge import BaseCharge
+from expungeservice.models.charge_types.charge import Charge
 from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
 
 
-class FelonyClassB(BaseCharge):
+class FelonyClassB(Charge):
 
     def __init__(self, **kwargs):
         super(FelonyClassB, self).__init__(**kwargs)
