+++ src/backend/expungeservice/models/charge_types/felony_class_c.py
@@ -1,8 +1,8 @@
-from expungeservice.models.charge_types.base_charge import BaseCharge
+from expungeservice.models.charge_types.charge import Charge
 from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
 
 
-class FelonyClassC(BaseCharge):
+class FelonyClassC(Charge):
 
     def __init__(self, **kwargs):
         super(FelonyClassC, self).__init__(**kwargs)
