+++ src/backend/expungeservice/models/charge_types/juvenile_charge.py
@@ -1,8 +1,8 @@
-from expungeservice.models.charge_types.base_charge import BaseCharge
+from expungeservice.models.charge_types.charge import Charge
 from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
 
 
-class JuvenileCharge(BaseCharge):
+class JuvenileCharge(Charge):
 
     def __init__(self, **kwargs):
         super(JuvenileCharge, self).__init__(**kwargs)
