+++ src/backend/expungeservice/models/charge_types/list_b.py
@@ -1,7 +1,7 @@
-from expungeservice.models.charge_types.base_charge import BaseCharge
+from expungeservice.models.charge_types.charge import Charge
 from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
 
-class ListB(BaseCharge):
+class ListB(Charge):
 
     def __init__(self, **kwargs):
         super(ListB, self).__init__(**kwargs)
