+++ src/backend/expungeservice/models/charge_types/parking_ticket.py
@@ -1,8 +1,8 @@
-from expungeservice.models.charge_types.base_charge import BaseCharge
+from expungeservice.models.charge_types.charge import Charge
 from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
 
 
-class ParkingTicket(BaseCharge):
+class ParkingTicket(Charge):
 
     def __init__(self, **kwargs):
         super(ParkingTicket, self).__init__(**kwargs)
