+++ src/backend/expungeservice/models/charge_types/person_crime.py
@@ -1,8 +1,8 @@
-from expungeservice.models.charge_types.base_charge import BaseCharge
+from expungeservice.models.charge_types.charge import Charge
 from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
 
 
-class PersonCrime(BaseCharge):
+class PersonCrime(Charge):
 
     def __init__(self, **kwargs):
         super(PersonCrime, self).__init__(**kwargs)
